updating wager process and migration tables

+++ app/database/migrations/2014_02_10_181700_create_all_tables.php
@@ -94,9 +94,9 @@ class CreateAllTables extends Migration {
 			$table->increments('id');
 			$table->integer('user_id')->unsigned();
 			$table->integer('session_id');
-			$table->integer('wagerUnitValue');
-			$table->integer('wagerTotalValue');
-			$table->integer('pointsLost')->default(0);
+			$table->float('wagerUnitValue');
+			$table->float('wagerTotalValue');
+			$table->float('pointsLost')->default(0);
 			$table->timestamps();
 		});
